wallet: move connect_missing_block next to its only caller
